requirements-dev.txt
pytest==7.2.0
PyYAML==6.0
dictdiffer==0.9.0
